AL-3 factory method retirado
No se ve necesidad de usar factory method para crear antenas (por ahora). Flake8 para linting configurado.
