+++ .flake8
@@ -1,0 +1,5 @@
+[flake8]
+exclude =
+    venv
+    __pycache__
+max-line-length = 120
