+++ requirements.dev.txt
@@ -1,1 +1,2 @@
 pytest~=7.4.0
+flake8~=6.1.0
