+++ src/array_antennas/__init__.py
@@ -1,0 +1,3 @@
+from .array import AntennaArray
+
+__all__ = [AntennaArray]
